requirements.txt
deepface
fastapi
uvicorn
pyjwt
python-multipart
python-dotenv
redis
tf-keras
Faker